+++ tests/testthat.R
@@ -1,4 +1,3 @@
-library(prettycheck)
 library(testthat)
 
 library(groomr)