+++ NEWS.md
@@ -1,7 +1,8 @@
-# groomr 0.1.0.9000 (Development Version)
+# groomr (Development Version)
 
 - Added `get_file_sizes()` function to get the length files.
 - Added `replace_caps()` function to replace caps in an atomic vector.
+- Added `parse_html_table()` function to parse HTML tables from a `character` vector.
 
 # groomr 0.1.0